Moved the version endpoint into its own controller to observe best practice and enhance testability.
